build(Makefile): change build directory and add deploy targets
Change the build directory from `_build` to `../../SPPDs-Bib-Docs`
to align with the new documentation structure. This change ensures
that the documentation is built in the correct location.

Add a new `deploy` target to automate the process of committing
documentation updates. This target navigates to the destination
path, stages changes, and commits them with a predefined message.

Introduce a `html-and-deploy` target that combines the `html` and
`deploy` targets, streamlining the process of building and deploying
documentation in a single command. This enhances the efficiency of
the documentation workflow.