feat(docs): add audio file and embed audio player in salins:25:solutions documentation
Add a new audio file `salins-25-solutions.wav` to the documentation
and embed an audio player in the `salins:25:solutions.rst` file. This
enhances the documentation by providing an audio version of the content,
making it more accessible to users who prefer listening over reading.
Additionally, a download link is provided for users who wish to save
the audio file locally.