feat(docs): add audio file and player for tracy.widom:01:on entry
Include an audio file and player in the documentation for the
tracy.widom:01:on entry. This addition enhances accessibility and
engagement by providing an audio version of the content. The audio
player is embedded in the HTML, allowing users to listen directly
from the page. A download link is also provided for convenience.