feat(docs): add audio file and player for sina:92:statistics entry
Add a new audio file 'sina-92-statistics.wav' to the documentation
and include an audio player in the 'sina:92:statistics.rst' file.
This enhancement provides an auditory option for users who prefer
listening to the content. The audio player is embedded using HTML,
and a download link is provided for convenience. This change aims
to improve accessibility and user experience by offering multiple
ways to consume the information.